Move common mpp code to SyscallHandler interface
