+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/input/KeyboardInput.kt
@@ -1,17 +1,16 @@
 package com.github.ajalt.mordant.input
 
-import com.github.ajalt.mordant.internal.enterRawModeMpp
-import com.github.ajalt.mordant.internal.readKeyMpp
+import com.github.ajalt.mordant.internal.SYSCALL_HANDLER
 import com.github.ajalt.mordant.terminal.Terminal
 import kotlin.time.Duration
 
-// TODO docs
+// TODO docs, tests
 fun Terminal.readKey(timeout: Duration = Duration.INFINITE): KeyboardEvent? {
     if (!info.inputInteractive) return null
-    return readKeyMpp(timeout)
+    return SYSCALL_HANDLER.readKeyEvent(timeout)
 }
 
 fun Terminal.enterRawMode(): AutoCloseable {
     if (!info.inputInteractive) return AutoCloseable { }
-    return enterRawModeMpp()
+    return SYSCALL_HANDLER.enterRawMode()
 }
