+++ mordant/src/jvmMain/kotlin/com/github/ajalt/mordant/internal/syscalls/SyscallHandler.jna.macos.kt
@@ -1,14 +1,10 @@
-package com.github.ajalt.mordant.internal.jna
+package com.github.ajalt.mordant.internal.syscalls
 
-import com.github.ajalt.mordant.input.internal.PosixRawModeHandler
-import com.github.ajalt.mordant.internal.MppImpls
 import com.github.ajalt.mordant.internal.Size
 import com.oracle.svm.core.annotate.Delete
 import com.sun.jna.*
 import java.io.IOException
 import java.util.concurrent.TimeUnit
-
-private const val TCSANOW: Int = 0x0
 
 @Delete
 @Suppress("ClassName", "PropertyName", "MemberVisibilityCanBePrivate", "SpellCheckingInspection")
@@ -73,27 +69,19 @@ private interface MacosLibC : Library {
 }
 
 @Delete
-internal class JnaMacosMppImpls : MppImpls, PosixRawModeHandler() {
-    @Suppress("SpellCheckingInspection")
-    private companion object {
-        const val STDIN_FILENO = 0
-        const val STDOUT_FILENO = 1
-        const val STDERR_FILENO = 2
-
-        val TIOCGWINSZ = when {
-            Platform.isMIPS() || Platform.isPPC() || Platform.isSPARC() -> 0x40087468L
-            else -> 0x00005413L
-        }
+@Suppress("SpellCheckingInspection")
+internal object SyscallHandlerJnaMacos : SyscallHandlerJnaPosix() {
+    private const val TCSANOW: Int = 0x0
+    private val TIOCGWINSZ = when {
+        Platform.isMIPS() || Platform.isPPC() || Platform.isSPARC() -> 0x40087468L
+        else -> 0x00005413L
     }
 
     private val libC: MacosLibC = Native.load(Platform.C_LIBRARY_NAME, MacosLibC::class.java)
-    override fun stdoutInteractive(): Boolean = libC.isatty(STDOUT_FILENO) == 1
-    override fun stdinInteractive(): Boolean = libC.isatty(STDIN_FILENO) == 1
-    override fun stderrInteractive(): Boolean = libC.isatty(STDERR_FILENO) == 1
-
+    override fun isatty(fd: Int): Int = libC.isatty(fd)
     override fun fastIsTty(): Boolean = false
     override fun getTerminalSize(): Size? {
-        // TODO: this seems to fail on macosArm64, use stty on mac for now
+        // TODO: JNA has a bug that causes this to fail on macosArm64, use stty on mac for now
 //        val size = MacosLibC.winsize()
 //        return if (libC.ioctl(STDIN_FILENO, NativeLong(TIOCGWINSZ), size) < 0) {
 //            null
