+++ mordant/src/linuxMain/kotlin/com/github/ajalt/mordant/internal/MppInternal.linux.kt
@@ -1,24 +1,3 @@
-@file:OptIn(ExperimentalForeignApi::class)
-
 package com.github.ajalt.mordant.internal
-
-import kotlinx.cinterop.ExperimentalForeignApi
-import kotlinx.cinterop.alloc
-import kotlinx.cinterop.memScoped
-import platform.posix.STDIN_FILENO
-import platform.posix.TIOCGWINSZ
-import platform.posix.ioctl
-import platform.posix.winsize
-
-internal actual fun getTerminalSize(): Size? {
-    return memScoped {
-        val size = alloc<winsize>()
-        if (ioctl(STDIN_FILENO, TIOCGWINSZ.toULong(), size) < 0) {
-            null
-        } else {
-            Size(width = size.ws_col.toInt(), height = size.ws_row.toInt())
-        }
-    }
-}
 
 internal actual fun hasFileSystem(): Boolean = true
